Remove top bar, add new logo, and put in mobile-friendly menu
Removes the top bar, replacing it with a single bar that is either on
the side (wide display) or the top (narrow display: mobile view).

Adds arrows to show the presence of submenus.  Adds Javascript to
expand submenus without opening them.

In the mobile view, a hamburger menu button is used.  Javascript
toggles the menu visibility when it is clicked.

Adds a new logo based on a Dr. Memory logo.